+++ nautilus_core/model/src/ffi/identifiers/order_list_id.rs
@@ -13,8 +13,23 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod data;
-pub mod events;
-pub mod identifiers;
-pub mod orderbook;
-pub mod types;
+use std::ffi::c_char;
+
+use nautilus_core::ffi::string::cstr_to_str;
+
+use crate::identifiers::order_list_id::OrderListId;
+
+/// Returns a Nautilus identifier from a C string pointer.
+///
+/// # Safety
+///
+/// - Assumes `ptr` is a valid C string pointer.
+#[no_mangle]
+pub unsafe extern "C" fn order_list_id_new(ptr: *const c_char) -> OrderListId {
+    OrderListId::from(cstr_to_str(ptr))
+}
+
+#[no_mangle]
+pub extern "C" fn order_list_id_hash(id: &OrderListId) -> u64 {
+    id.value.precomputed_hash()
+}
