+++ nautilus_core/model/src/ffi/identifiers/symbol.rs
@@ -13,8 +13,23 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod data;
-pub mod events;
-pub mod identifiers;
-pub mod orderbook;
-pub mod types;
+use std::ffi::c_char;
+
+use nautilus_core::ffi::string::cstr_to_str;
+
+use crate::identifiers::symbol::Symbol;
+
+/// Returns a Nautilus identifier from a C string pointer.
+///
+/// # Safety
+///
+/// - Assumes `ptr` is a valid C string pointer.
+#[no_mangle]
+pub unsafe extern "C" fn symbol_new(ptr: *const c_char) -> Symbol {
+    Symbol::from(cstr_to_str(ptr))
+}
+
+#[no_mangle]
+pub extern "C" fn symbol_hash(id: &Symbol) -> u64 {
+    id.value.precomputed_hash()
+}
