+++ nautilus_core/model/src/ffi/orderbook/mod.rs
@@ -13,8 +13,5 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod data;
-pub mod events;
-pub mod identifiers;
-pub mod orderbook;
-pub mod types;
+pub mod book;
+pub mod level;
