+++ nautilus_core/model/src/orderbook/mod.rs
@@ -14,9 +14,5 @@
 // -------------------------------------------------------------------------------------------------
 
 pub mod book;
-#[cfg(feature = "ffi")]
-pub mod book_api;
 pub mod ladder;
 pub mod level;
-#[cfg(feature = "ffi")]
-pub mod level_api;
