+++ nautilus_core/model/src/python/identifiers/mod.rs
@@ -13,8 +13,4 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-pub mod data;
-pub mod events;
-pub mod identifiers;
-pub mod orderbook;
-pub mod types;
+pub mod instrument_id;
